test(eth): group eth_{call,estimateGas} tests, expand skip-sender estimates
