+++ itests/fevm_test.go
@@ -556,105 +556,6 @@ func TestFEVMRecursiveActorCall(t *testing.T) {
 }
 
 // TestFEVMRecursiveActorCallEstimate
-func TestFEVMRecursiveActorCallEstimate(t *testing.T) {
-	ctx, cancel, client := kit.SetupFEVMTest(t)
-	defer cancel()
-
-	// install contract Actor
-	filenameActor := "contracts/ExternalRecursiveCallSimple.hex"
-	_, actorAddr := client.EVM().DeployContractFromFilename(ctx, filenameActor)
-
-	contractAddr, err := ethtypes.EthAddressFromFilecoinAddress(actorAddr)
-	require.NoError(t, err)
-
-	// create a new Ethereum account
-	key, ethAddr, ethFilAddr := client.EVM().NewAccount()
-	kit.SendFunds(ctx, t, client, ethFilAddr, types.FromFil(1000))
-
-	makeParams := func(r int) []byte {
-		funcSignature := "exec1(uint256)"
-		entryPoint := kit.CalcFuncSignature(funcSignature)
-
-		inputData := make([]byte, 32)
-		binary.BigEndian.PutUint64(inputData[24:], uint64(r))
-
-		params := append(entryPoint, inputData...)
-
-		return params
-	}
-
-	testN := func(r int) func(t *testing.T) {
-		return func(t *testing.T) {
-			t.Logf("running with %d recursive calls", r)
-
-			params := makeParams(r)
-
-			gasParams, err := json.Marshal(ethtypes.EthEstimateGasParams{Tx: ethtypes.EthCall{
-				From: &ethAddr,
-				To:   &contractAddr,
-				Data: params,
-			}})
-			require.NoError(t, err)
-
-			gaslimit, err := client.EthEstimateGas(ctx, gasParams)
-			require.NoError(t, err)
-			require.LessOrEqual(t, int64(gaslimit), buildconstants.BlockGasLimit)
-
-			t.Logf("EthEstimateGas GasLimit=%d", gaslimit)
-
-			maxPriorityFeePerGas, err := client.EthMaxPriorityFeePerGas(ctx)
-			require.NoError(t, err)
-
-			nonce, err := client.MpoolGetNonce(ctx, ethFilAddr)
-			require.NoError(t, err)
-
-			tx := &ethtypes.Eth1559TxArgs{
-				ChainID:              buildconstants.Eip155ChainId,
-				To:                   &contractAddr,
-				Value:                big.Zero(),
-				Nonce:                int(nonce),
-				MaxFeePerGas:         types.NanoFil,
-				MaxPriorityFeePerGas: big.Int(maxPriorityFeePerGas),
-				GasLimit:             int(gaslimit),
-				Input:                params,
-				V:                    big.Zero(),
-				R:                    big.Zero(),
-				S:                    big.Zero(),
-			}
-
-			client.EVM().SignTransaction(tx, key.PrivateKey)
-			hash := client.EVM().SubmitTransaction(ctx, tx)
-
-			smsg, err := ethtypes.ToSignedFilecoinMessage(tx)
-			require.NoError(t, err)
-
-			_, err = client.StateWaitMsg(ctx, smsg.Cid(), 0, 0, false)
-			require.NoError(t, err)
-
-			receipt, err := client.EthGetTransactionReceipt(ctx, hash)
-			require.NoError(t, err)
-			require.NotNil(t, receipt)
-
-			t.Logf("Receipt GasUsed=%d", receipt.GasUsed)
-			t.Logf("Ratio %0.2f", float64(receipt.GasUsed)/float64(gaslimit))
-			t.Logf("Overestimate %0.2f", ((float64(gaslimit)/float64(receipt.GasUsed))-1)*100)
-
-			require.EqualValues(t, ethtypes.EthUint64(1), receipt.Status)
-		}
-	}
-
-	t.Run("n=1", testN(1))
-	t.Run("n=2", testN(2))
-	t.Run("n=3", testN(3))
-	t.Run("n=4", testN(4))
-	t.Run("n=5", testN(5))
-	t.Run("n=10", testN(10))
-	t.Run("n=20", testN(20))
-	t.Run("n=30", testN(30))
-	t.Run("n=40", testN(40))
-	t.Run("n=50", testN(50))
-	t.Run("n=100", testN(100))
-}
 
 // TestFEVMDeployWithValue deploys a contract while sending value to it
 func TestFEVMDeployWithValue(t *testing.T) {
@@ -1487,213 +1388,6 @@ func TestEthGetTransactionByBlockHashAndIndexAndNumber(t *testing.T) {
 		require.Error(t, err)
 		require.ErrorContains(t, err, "out of range")
 	})
-}
-
-func TestEthCall(t *testing.T) {
-	ctx, cancel, client := kit.SetupFEVMTest(t)
-	defer cancel()
-
-	filename := "contracts/Errors.hex"
-	fromAddr, contractAddr := client.EVM().DeployContractFromFilename(ctx, filename)
-
-	divideByZeroSignature := kit.CalcFuncSignature("failDivZero()")
-
-	_, _, err := client.EVM().InvokeContractByFuncName(ctx, fromAddr, contractAddr, "failDivZero()", []byte{})
-	require.Error(t, err)
-
-	latestBlock, err := client.EthBlockNumber(ctx)
-	require.NoError(t, err)
-
-	contractAddrEth, err := ethtypes.EthAddressFromFilecoinAddress(contractAddr)
-	require.NoError(t, err)
-
-	callParams := ethtypes.EthCall{
-		From: nil,
-		To:   &contractAddrEth,
-		Data: divideByZeroSignature,
-	}
-
-	t.Run("FailedToProcessBlockParam", func(t *testing.T) {
-		invalidBlockNumber := latestBlock + 1000
-		_, err = client.EthCall(ctx, callParams, ethtypes.NewEthBlockNumberOrHashFromNumber(invalidBlockNumber))
-		require.Error(t, err)
-		require.Contains(t, err.Error(), "requested a future epoch (beyond 'latest')")
-	})
-
-	t.Run("DivideByZeroError", func(t *testing.T) {
-		_, err = client.EthCall(ctx, callParams, ethtypes.NewEthBlockNumberOrHashFromNumber(latestBlock))
-		require.Error(t, err)
-
-		var dataErr *api.ErrExecutionReverted
-		require.ErrorAs(t, err, &dataErr, "Expected error to be ErrExecutionReverted")
-		require.Regexp(t, `message execution failed [\s\S]+\[DivideByZero\(\)\]`, dataErr.Message)
-
-		// Get the error data
-		require.Equal(t, dataErr.Data, "0x4e487b710000000000000000000000000000000000000000000000000000000000000012", "Expected error data to contain 'DivideByZero()'")
-	})
-}
-
-func TestEthEstimateGas(t *testing.T) {
-	ctx, cancel, client := kit.SetupFEVMTest(t)
-	defer cancel()
-
-	_, ethAddr, filAddr := client.EVM().NewAccount()
-
-	kit.SendFunds(ctx, t, client, filAddr, types.FromFil(10))
-
-	filename := "contracts/Errors.hex"
-	_, contractAddr := client.EVM().DeployContractFromFilename(ctx, filename)
-
-	contractAddrEth, err := ethtypes.EthAddressFromFilecoinAddress(contractAddr)
-	require.NoError(t, err)
-
-	testCases := []struct {
-		name           string
-		function       string
-		expectedError  string
-		expectedErrMsg interface{}
-	}{
-		{"DivideByZero", "failDivZero()", "0x4e487b710000000000000000000000000000000000000000000000000000000000000012", `message execution failed [\s\S]+\[DivideByZero\(\)\]`},
-		{"Assert", "failAssert()", "0x4e487b710000000000000000000000000000000000000000000000000000000000000001", `message execution failed [\s\S]+\[Assert\(\)\]`},
-		{"RevertWithReason", "failRevertReason()", fmt.Sprintf("%x", []byte("my reason")), `message execution failed [\s\S]+\[Error\(my reason\)\]`},
-		{"RevertEmpty", "failRevertEmpty()", "0x", `message execution failed [\s\S]+\[none\]`},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			signature := kit.CalcFuncSignature(tc.function)
-
-			callParams := ethtypes.EthCall{
-				From: &ethAddr,
-				To:   &contractAddrEth,
-				Data: signature,
-			}
-
-			gasParams, err := json.Marshal(ethtypes.EthEstimateGasParams{Tx: callParams})
-			require.NoError(t, err, "Error marshaling gas params")
-
-			_, err = client.EthEstimateGas(ctx, gasParams)
-
-			if tc.expectedError != "" {
-				require.Error(t, err)
-				var dataErr *api.ErrExecutionReverted
-				require.ErrorAs(t, err, &dataErr, "Expected error to be ErrExecutionReverted")
-				require.Regexp(t, tc.expectedErrMsg, dataErr.Message)
-				require.Contains(t, dataErr.Data, tc.expectedError)
-			}
-		})
-	}
-}
-
-func TestEthNullRoundHandling(t *testing.T) {
-	blockTime := 100 * time.Millisecond
-	client, _, ens := kit.EnsembleMinimal(t, kit.MockProofs(), kit.ThroughRPC())
-
-	bms := ens.InterconnectAll().BeginMining(blockTime)
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Minute)
-	defer cancel()
-
-	client.WaitTillChain(ctx, kit.HeightAtLeast(10))
-
-	bms[0].InjectNulls(10)
-
-	tctx, cancel := context.WithTimeout(ctx, 30*time.Second)
-	defer cancel()
-	ch, err := client.ChainNotify(tctx)
-	require.NoError(t, err)
-	<-ch
-	hc := <-ch
-	require.Equal(t, store.HCApply, hc[0].Type)
-
-	afterNullHeight := hc[0].Val.Height()
-
-	nullHeight := afterNullHeight - 1
-	for nullHeight > 0 {
-		ts, err := client.ChainGetTipSetByHeight(ctx, nullHeight, types.EmptyTSK)
-		require.NoError(t, err)
-		if ts.Height() == nullHeight {
-			nullHeight--
-		} else {
-			break
-		}
-	}
-
-	nullBlockHex := fmt.Sprintf("0x%x", int(nullHeight))
-	nullBlockParam := ethtypes.NewEthBlockNumberOrHashFromNumber(ethtypes.EthUint64(nullHeight))
-	client.WaitTillChain(ctx, kit.HeightAtLeast(nullHeight+2))
-	testCases := []struct {
-		name     string
-		testFunc func() error
-	}{
-		{
-			name: "EthGetBlockByNumber",
-			testFunc: func() error {
-				_, err := client.EthGetBlockByNumber(ctx, nullBlockHex, true)
-				return err
-			},
-		},
-		{
-			name: "EthGetBlockTransactionCountByNumber",
-			testFunc: func() error {
-				_, err := client.EthGetBlockTransactionCountByNumber(ctx, nullBlockHex)
-				return err
-			},
-		},
-		{
-			name: "EthGetTransactionByBlockNumberAndIndex",
-			testFunc: func() error {
-				_, err := client.EthGetTransactionByBlockNumberAndIndex(ctx, nullBlockHex, ethtypes.EthUint64(0))
-				return err
-			},
-		},
-		{
-			name: "EthGetBlockReceipts",
-			testFunc: func() error {
-				_, err := client.EthGetBlockReceipts(ctx, nullBlockParam)
-				return err
-			},
-		},
-		{
-			name: "EthGetBlockReceiptsLimited",
-			testFunc: func() error {
-				_, err := client.EthGetBlockReceiptsLimited(ctx, nullBlockParam, api.LookbackNoLimit)
-				return err
-			},
-		},
-		{
-			name: "EthTraceBlock",
-			testFunc: func() error {
-				_, err := client.EthTraceBlock(ctx, nullBlockHex)
-				return err
-			},
-		},
-		{
-			name: "EthTraceReplayBlockTransactions",
-			testFunc: func() error {
-				_, err := client.EthTraceReplayBlockTransactions(ctx, nullBlockHex, []string{"trace"})
-				return err
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := tc.testFunc()
-			require.Error(t, err)
-
-			// Test errors.Is
-			require.ErrorIs(t, err, new(api.ErrNullRound), "error should be or wrap ErrNullRound")
-
-			// Test errors.As and verify message
-			var nullRoundErr *api.ErrNullRound
-			require.ErrorAs(t, err, &nullRoundErr, "error should be convertible to ErrNullRound")
-
-			expectedMsg := fmt.Sprintf("requested epoch was a null round (%d)", nullHeight)
-			require.Equal(t, expectedMsg, nullRoundErr.Error())
-			require.Equal(t, nullHeight, nullRoundErr.Epoch)
-		})
-	}
 }
 
 func TestFEVMEamCreateTwiceFail(t *testing.T) {
